[MeL] Move ElementErrorCode from MeshQuality to Elements.
